+++ src/edu/umass/cs/gns/nsdesign/recordmap/ReplicaControllerRecord.java
@@ -361,9 +361,9 @@ public class ReplicaControllerRecord {
    *
    * @return
    */
-  public int getActiveVersion() throws FieldNotFoundException {
+  public short getActiveVersion() throws FieldNotFoundException {
     if (hashMap.containsKey(ACTIVE_VERSION)) {
-      return (Integer) hashMap.get(ACTIVE_VERSION);
+      return (Short) hashMap.get(ACTIVE_VERSION);
     }
     throw new FieldNotFoundException(ACTIVE_VERSION);
   }
@@ -373,9 +373,9 @@ public class ReplicaControllerRecord {
    *
    * @return
    */
-  public int getOldActiveVersion() throws FieldNotFoundException {
+  public short getOldActiveVersion() throws FieldNotFoundException {
     if (hashMap.containsKey(OLD_ACTIVE_VERSION)) {
-      return (Integer) hashMap.get(OLD_ACTIVE_VERSION);
+      return (Short) hashMap.get(OLD_ACTIVE_VERSION);
     }
     throw new FieldNotFoundException(OLD_ACTIVE_VERSION);
   }
